mutants/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurora-trader"
version = "0.1.0"
description = "Aurora Trading System"
authors = [{name = "Jennifer", email = "jennifer@aurora.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.0",
    "pyyaml>=6.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "mutmut>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "bandit>=1.7.0",
    "pip-audit>=1.0.0",
]

[project.scripts]
trading-bot = "enhanced_paper_trading:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "scripts*", "tests*", "utils*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["core", "strategies", "features", "brokers"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008", "B006"]

[tool.ruff.lint]
extend-select = ["E","F","I","B","UP","SIM"]
ignore = ["E203","E266"]
exclude = [
    "attic/",
    "docs/",
    "tests/",
    "*.md",
    "*.txt",
    "*.json",
    "*.yaml",
    "*.yml"
]

[tool.ruff.format]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
]

[tool.mutmut]
paths_to_mutate = ["core/data_sanity"]
tests_dir = "tests"
runner = "python -m pytest tests/ -q"
backup = false
CI = true
dict_synonyms = ["Struct", "NamedStruct"]
